Add posts

+++ src/users/routers.py
@@ -1,7 +1,5 @@
-from typing import List
-
 from fastapi import APIRouter, HTTPException, Query, Response, status
-from sqlmodel import asc, select
+from sqlmodel import desc, select
 
 from config.auth import auth_dep, decode_access_token
 from config.db import session_dep
@@ -17,35 +15,39 @@ async def get_users(
     session: session_dep,
     offset: int = Query(0, ge=0),
     limit: int = Query(10, ge=1, le=100),
-) -> List[UserRead]:
+) -> list[UserRead]:
     try:
-        statement = select(User).order_by(asc(User.id)).offset(offset).limit(limit)
+        statement = (
+            select(User).order_by(desc(User.created_at)).offset(offset).limit(limit)
+        )
         results = await session.execute(statement)
         rows = results.scalars().all()
         return rows
     except HTTPException:
         raise
     except Exception as e:
-        logger.debug(f"ERROR: {e}")
+        logger.error(f"ERROR: {e}")
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
 @users_router.delete("/{user_id}")
 async def delete_user(session: session_dep, auth: auth_dep, user_id: int) -> Response:
     try:
-        if decode_access_token(auth) != user_id:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
-
         statement = select(User).where(User.id == user_id)
         results = await session.execute(statement)
         row = results.scalar()
 
+        if not row:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
+        if decode_access_token(auth) != user_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
         await session.delete(row)
         await session.commit()
-
-        return Response(status_code=status.HTTP_200_OK)
+        return Response(status_code=status.HTTP_204_NO_CONTENT)
     except HTTPException:
         raise
     except Exception as e:
-        logger.debug(f"ERROR: {e}")
+        logger.error(f"ERROR: {e}")
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
